Renamed FloatImpl to SingleImpl.
